修改项： 个人id为int，但传入了long 创建订单失败 订单取消失败 修改购物车内的商品 修改购物车商品数量

+++ src/main/java/org/taobao/controller/UserController.java
@@ -158,7 +158,7 @@ public class UserController {
             userProfileVO.setUpdateTime(user.getUpdateTime());
 
             // 获取订单状态统计
-            Map<String, Long> orderStats = orderService.getOrderStatusStatistics(user.getUserId());
+            Map<String, Long> orderStats = orderService.getOrderStatusStatistics(user.getUserId().intValue());
             // 设置订单状态统计到VO中
             userProfileVO.setPendingOrderCount(orderStats.getOrDefault("pending", 0L));
             userProfileVO.setPaidOrderCount(orderStats.getOrDefault("paid", 0L));
